Relocate oep functions
