+++ scripts/upload_results_to_oep.py
@@ -31,7 +31,6 @@ The oemetadata format is a standardised json file format and is required for all
 the OEP. It includes the data model, the used data types, and general information about the data
 context. Tables in sqlalchemy are created based on the information in the oemetadata.
 """
-import json
 import logging
 import os
 import pathlib
@@ -41,8 +40,7 @@ from datetime import date
 import pandas as pd
 
 from oemof_b3.config import config
-from oemof_b3.schema import oemetadata_scal, oemetadata_ts, SCHEMA_SCAL, SCHEMA_TS
-from oemof_b3.tools import data_processing as dp
+from oemof_b3.tools.oep import create_metadata, save_dict_to_json
 
 try:
     from oem2orm import oep_oedialect_oem2orm as oem2orm
@@ -62,49 +60,9 @@ except AttributeError:
     )
 
 
-def save_dict_to_json(data, filepath, encoding="utf-8"):
-    with open(filepath, "w", encoding=encoding) as f:
-        return json.dump(data, f, sort_keys=False, indent=2)
-
-
-def list_diff(sample, default):
-    r"""
-    Determines extra and missing items in sample in
-    comparison with default.
-
-    Parameters
-    ----------
-    sample : list
-    default : list
-
-    Returns
-    -------
-    (extra_items, missing_items): (list, list)
-    """
-    extra_items = dp.get_list_diff(sample, default)
-
-    missing_items = dp.get_list_diff(default, sample)
-
-    if not extra_items and not missing_items:
-        return None
-
-    else:
-        return (extra_items, missing_items)
-
-
-def create_metadata(data, template=None):
-    if template is not None:
-        pass
-    elif list_diff(data.columns, SCHEMA_SCAL.columns) is None:
-        template = oemetadata_scal
-    elif list_diff(data.columns, SCHEMA_TS.columns) is None:
-        template = oemetadata_ts
-    else:
-        raise ValueError("Could not match data with the existing templates.")
-
-    metadata = template.copy()
-
-    return metadata
+# define table names
+def get_table_name(filename, scenario):
+    return f"{os.path.splitext(filename)[0]}_{scenario}"
 
 
 def write_metadata(metadata, schema, table, title, keywords):
@@ -133,10 +91,6 @@ if __name__ == "__main__":
 
     if not os.path.exists(metadata_path):
         os.makedirs(metadata_path)
-
-    # define table names
-    def get_table_name(filename, scenario):
-        return f"{os.path.splitext(filename)[0]}_{scenario}"
 
     # find data to upload
     dict_table_filename = {
